% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{add_group_sub}
\alias{add_group_sub}
\title{Add sub index within group}
\usage{
add_group_sub(df, group, sep = "_", tmp_col = "tmp_col")
}
\arguments{
\item{df}{A dataframe.}

\item{group}{A string or string vector.
When vector, the first string will be used for
adding sub index.}

\item{sep}{A string for separator.}

\item{tmp_col}{A string of colnames for temporary use.}
}
\value{
A dataframe.
}
\description{
Add sub index within group
}
\examples{
library(dplyr)
add_group_sub(mtcars, c("am", "gear"))
add_group_sub(mtcars, c("cyl", "am"))

}
