% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_col}
\alias{extract_col}
\title{Helper for na_col_omit()}
\usage{
extract_col(col, df)
}
\arguments{
\item{col}{A string or string vector.}

\item{df}{A dataframe.}
}
\value{
A vector.
}
\description{
Helper for na_col_omit()
}
\examples{
library(tidyr)
library(dplyr)
library(purrr)
library(ggplot2)
hogwarts \%>\%
  pivot(row = "hour", col = "wday", 
        value = c("subject", "teacher", "room"), 
        split = c("house", "grade"))
hogwarts \%>\%
  pivot(row = "hour", col = "wday", 
        value = c("subject", "room", "house", "grade"), 
        split = c("teacher"))
starwars \%>\%
  pivot(row = "homeworld", col = "species", value = "name", split = "sex")
msleep \%>\%
  pivot(row = "vore", col = "conservation", value = "name") \%>\%
  na2empty() \%>\%
  print(n = nrow(.))
tibble::as_tibble(Titanic) \%>\%
  pivot(row = "Age", col = c("Sex", "Survived"), 
        value = "n", split = "Class")
diamonds \%>\%
  pivot(row = "cut", col = "color", value = "price", split = "clarity")

}
