% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_col}
\alias{has_col}
\title{Detect if df has col}
\usage{
has_col(df, col)
}
\arguments{
\item{df}{A dataframe.}

\item{col}{A string or string vector.}
}
\value{
A dataframe.
}
\description{
Detect if df has col
}
\examples{
colnames(mtcars)
has_col(mtcars, c("mpg", "cyl"))
has_col(mtcars, c("mpg", "foo"))

}
