% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hogwarts}
\alias{hogwarts}
\title{Timetable in Hogwarts School of Witchcraft and Wizardry.}
\format{
A data frame with 548 rows and 7 variable:
\describe{
\item{grade}{Grades in school.}
\item{house}{Houses. G: Gryffindor, S: Slytherin, R: Ravenclaw, and H: Hufflepuff.}
\item{wday}{Abbreviations of day of the week.}
\item{hour}{Hours.}
\item{teacher}{Teachers.}
\item{subject}{Subjects.}
\item{room}{}
}
}
\usage{
hogwarts
}
\description{
Timetable in Hogwarts School of Witchcraft and Wizardry.
}
\examples{
data(hogwarts)
hogwarts
}
\keyword{datasets}
