% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{omit_na_rows}
\alias{omit_na_rows}
\title{Remove all NA rows}
\usage{
omit_na_rows(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Remove all NA rows
}
\examples{
library(tidyr)
library(dplyr)
library(purrr)
library(ggplot2)
starwars \%>\%
  pivot(row = "homeworld", col = "species", value = "name", split = "sex")
msleep \%>\%
  pivot(row = "vore", col = "conservation", value = "name") \%>\%
  na2empty() \%>\%
  print(n = nrow(.))
as_tibble(Titanic) \%>\%
  pivot(row = "Age", col = c("Sex", "Survived"), value = "n", split = "Class")
diamonds \%>\%
  pivot(row = "cut", col = "color", value = "price", split = "clarity")

}
