% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_col}
\alias{replace_col}
\title{Replace a col with a data.frame.}
\usage{
replace_col(df, replace)
}
\arguments{
\item{df, replace}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Replace a col with a data.frame.
}
\examples{
(state <- tibble::tibble(state = state.name, area = state.area))
(abb <- tibble::tibble(state = state.name, abb = state.abb))
replace_col(state, abb)

}
