% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{split_force}
\alias{split_force}
\title{Split by force with "" when split is NULL}
\usage{
split_force(df, split)
}
\arguments{
\item{df}{A dataframe.}

\item{split}{A string or string vector.}
}
\value{
A dataframe.
}
\description{
Split by force with "" when split is NULL
}
\examples{
split_force(mtcars, split = NULL)
split_force(mtcars, split = c("cyl"))

}
