% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{validate_col}
\alias{validate_col}
\title{Validate col}
\usage{
validate_col(df, col)
}
\arguments{
\item{df}{A dataframe.}

\item{col}{A string or string vector.}
}
\value{
A dataframe.
}
\description{
Validate col
}
\examples{
library(tidyr)
library(dplyr)
library(purrr)
library(ggplot2)
starwars \%>\%
  pivot(row = "homeworld", col = "species", value = "name", split = "sex")
msleep \%>\%
  pivot(row = "vore", col = "conservation", value = "name") \%>\%
  na2empty() \%>\%
  print(n = nrow(.))
as_tibble(Titanic) \%>\%
  pivot(row = "Age", col = c("Sex", "Survived"), value = "n", split = "Class")
diamonds \%>\%
  pivot(row = "cut", col = "color", value = "price", split = "clarity")

}
